// instituto-granada/src/pages/Volunteers/types.ts
export type FormTexts = {
  buttonLabel: string;
  emailLabel: string;
  emailPlaceholder: string;
  nameLabel: string;
  namePlaceholder: string;
  messageLabel: string;
  messageInputMaxLenght: string;
  messagePlaceholder: string;
  positionLabel: string;
  positionPlaceholder: string;
  subtitle: string;
  title: string;
};